\name{extract.rule}
\alias{extract.rule}

\title{An example function from the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R}

\description{An example function from Chapter 3 of the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. See Appendix B or http://www.wiley.com/go/data_mining_algorithms for more details.}

\usage{See Section 3.4, Example 3.4.3.}

\arguments{See Section 3.4, Example 3.4.3.}

\details{See Section 3.4, Example 3.4.3.}

\value{See Section 3.4, Example 3.4.3.}

\references{Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. Wiley.}

\author{
Pawel Cichosz <p.cichosz@elka.pw.edu.pl>
}

\note{
}


\seealso{
\code{\link{rpart}}
\code{\link{rewrite.splits}}
\code{\link{leaf.error}}
}

\examples{
}

\keyword{models}
